Rename debugger to debugging_util
